small warning fix

+++ BasicPNGWriter.cpp
@@ -107,8 +107,8 @@ uint8_t* BasicPNGWriter::getPixelClamped(int x, int y)
 {
 	if (x < 0) x = 0;
 	if (y < 0) y = 0;
-	if (x > m_width - 1) x = static_cast<int>(m_width - 1);
-	if (y > m_height - 1) y = static_cast<int>(m_height - 1);
+	if (x > static_cast<int>(m_width - 1)) x = static_cast<int>(m_width - 1);
+	if (y > static_cast<int>(m_height - 1)) y = static_cast<int>(m_height - 1);
 
 	return &m_buffer[x*CHANSPERPIXEL + y * (m_width * CHANSPERPIXEL)];
 }
